Move is_float function to utils.py
